requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0,<1.5.0
anthropic>=0.3.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.5,<0.1.0
requests>=2.26.0,<3.0.0
praw>=7.4.0,<8.0.0
python-dotenv>=0.19.0,<0.20.0
tenacity>=8.0.0,<9.0.0
alembic>=1.7.0,<1.8.0
slowapi==0.1.8